schedule summary refactor
